Nested types: expose table column types through TableDescriptor
Before we only had type information for each SlotDescriptor, rather
than for the entire table. However, in order to do schema resolution
for nested fields, we need to be able to traverse the table schema
starting from the table-level columns.

We could theoretically expose only the paths needed to resolve each
slot, but it's simpler to have the whole table.

Change-Id: I026c1f1f552d1ac5d1b267f876e1c39a258714b5
Reviewed-on: http://gerrit.cloudera.org:8080/404
Reviewed-by: Skye Wanderman-Milne <skye@cloudera.com>
Tested-by: Internal Jenkins