IMPALA-2266: Pass correct child node in 2nd phase merge aggregation.
The bug was that we were not passing the correct child node in the
constructor of the AggregationNode for the 2nd phase merge. As a result,
the 2nd phase merge aggregation node did not have the correct output
smap set, and all exprs on top were not substituted properly.

The bug was not apparent because the child node of the 2nd phase distinct
aggregation was fixed up after construction when setting the new plan
root of the corresponding fragment, so the bug did not manifest itself in
cases where the output smap was empty anyway (e.g., distinct aggregations
in the top-level select block).

Change-Id: I0617b2389e55ebcaf424600803babb250323ad3c
Reviewed-on: http://gerrit.cloudera.org:8080/715
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Internal Jenkins