Make TDdlType compatible with older clients
A disaster recovery application that uses thrift to directly execute
DDL (instead of using SQL) stopped working. A client call to drop a
function ended up down the code path to drop a view. Apparently
commit 47d061 messed up the enum ordering for old clients by adding
TRUNCATE_TABLE in the middle of the enum list. The fix is to move
TRUNCATE_TABLE to the end. Commit 47d061 was never released so this
there shouldn't be concern about breaking newer clients.

Change-Id: I79ebec65497077471a37e5712061c418403a336a
Reviewed-on: http://gerrit.cloudera.org:8080/899
Reviewed-by: Casey Ching <casey@cloudera.com>
Tested-by: Internal Jenkins