IMPALA-2444: Part 3: Parquet scanner perf improvements
This patch focuses on optimizing NextLevels() and ReadValue().

Convert NextLevels to a template version so that rep_levels is
only maintained if needed. Otherwise, it is always 0. This change
requires a bit of plumbing because NextLevels() is called in several
places.

Avoid pointer indirection in DictDecoder by adding Reset() method. This
avoids having to chase an additional pointer in the inner loop.

After this set of changes the regression is ~5%.

Change-Id: I73c68a7018cf558117037012339f5ec2cfc2c9ea
Reviewed-on: http://gerrit.cloudera.org:8080/1111
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins