IMPALA-2378: Part 1: reset of PAGG accumulates function contexts
When the PAGG node is reset it allocates new function contexts for the
aggregation functions. Previously, to ensure they were cleaned up,
it stashed them in a per-query object pool. Now that we reset nodes,
this meant that the per-query object pool grows every time the subplan
is reset. The fix is to delete the function contexts when the partition
is closed.

Change-Id: Ie3e4867e4de9e65883e015970b7cf1af618763d8
Reviewed-on: http://gerrit.cloudera.org:8080/1140
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Internal Jenkins