Update task_size to use total bytes of scan ranges
Currently the task_size is set to be the # of scan ranges.
And then, on the client side this is returned in InputSplit#getLength(),
which *should* an estimate of the # of bytes
for the split.

We should therefore change the task_size to
be the total # of bytes for all scan ranges in the task.

Change-Id: I55487c22282ac8ffe05d43992002138e6b8ea874
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/9006
Tested-by: jenkins
Reviewed-by: Li Li <lili@cloudera.com>