Add tmp table support for multiple planners
This adds support for tmp table in case where
there're multiple planners. Each planner now maintains a separete
temp table identified by planner's IP address followed by the process
ID, to avoid collision.

In addition, this adds support to bypass Hive Metastore for temp
table creation - it now only resides in memory cache. This is reasonable
because a temp table is only queried once immediately after it is
created.

All temp tables are created inside a temp database which can be customized
through flag. The name for this database should NOT collide with any
other databases in Hive Metastore.

All accesses on temp table for one planner should still be synchronized.
This is because path request consists of two steps: creating the temp
table, and querying the table (and could be more with partitioned table).
Without sync, these steps could be interleaved, and result could be incorrect.

Change-Id: Iee8016a50848ae4abaeff74e6042f9d261f33c24
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/9278
Tested-by: jenkins
Reviewed-by: Chao Sun <chao.sun@cloudera.com>