Populate ports when there are multiple recordserviced on one node
Currently, if there are multiple recordserviced processes running on
one node, we always pick the first one to resolve worker port. This
however doesn't work for the mini cluster tests. We would randomly
kill some process in the tests. But, since all locations for the
replicas have the same port, a job would immediately fail without
retrying. This is not ideal.

This patch changes it by assigning different ports the above locations
if they are available. This change should only affect the test
environment.

Change-Id: I84fad5e951f6d64729320aa94495462e27bb4abf
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/9137
Tested-by: jenkins
Reviewed-by: Chao Sun <chao.sun@cloudera.com>
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/9461
Reviewed-by: Li Li <lili@cloudera.com>