RS-94: Fix "Read timed out" error when RS server is busy
Where there are a large number of incoming requests during a short
time period, the number of connections could easily exceed the number
of worker threads in the Thrift server. Checking whether the former
equals to the latter is too restrictive, and will return "Read timed
out" error message to lots of those requests.

Change-Id: I2626af5e84ff6a0cad9d54518b9a5f3499ac2dbb
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/10370
Tested-by: jenkins
Reviewed-by: Lenni Kuff <lskuff@cloudera.com>