Change ZK to read Kerberos principal & keytab from command line parameter
Currently, when security is enabled, ZK reads the principal & keytab location
from the Hadoop configuration. This adds extra complexity since those info
is already passed to recordserviced as command line arguments. Also, in CM
deployment, we pick the configuration file under /etc/recordservice/conf, and
it is difficult to specify keytab location in this file (it's a bad idea
to copy keytab to this directory, and it's impossible to reference the one under
/var/run/cloudera-scm-agent/...).

Therefore, this patch uses the principal & keytab for securing ZK nodes. It also
changes how we set the default ACL for ZK nodes: if Kerberos is enabled, the user
(e.g., recordservice) who creates the Znode has full access permission. Otherwise,
the ACL is set to world:anyone:cdrwa.

Also, to enable planner discovery, in default we add an extra ACL: world:anyone:r
to the dir /recordservice/planners, so anyone can list the dir and get all the children
names. In this way, the client side apps (MR/Spark) do not need to use the same
principal & keytab the server is using in order to obtain planner membership. The
extra ACL can be customized by an additional property "recordservice.zookeeper.planners.acl".

This patch is manually tested on the 3-node cluster.

Change-Id: I3f31f66c58f454da8f19abfea033d379aee5b06a
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/11378
Tested-by: jenkins
Reviewed-by: Chao Sun <chao.sun@cloudera.com>