Move RecordService to the Impala virtualenv model
In July 2015 Impala made changes to execute its python code through
a virtual environment. This helps manage python packages in a cleaner
fashion than what we had been doing previously. RecordService, up to
this point had brought in some of that code but had not used it. The
patch begins the transition towards using that code.

This patch adds some missing packages and executes the hive serde
tests, the current RS tests that are located in the server tests/
directory, through this virtualenv.

Future work involves running the impala-shell through the virtualenv.
That will require more testing.

Change-Id: I5dc68bca57992e1c9178f38459a390c63e3b5719
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/11654
Tested-by: Jenkins User
Reviewed-by: Chao Sun <chao.sun@cloudera.com>