RS-140: Memory usage control for Compressed Text Files
This is a follow-up for dynamically adjust fetch size.

Currently, when dynamically adjusting the fetch size on server side, we cannot have
an accurate estimation for the row batch memory usage, especially for the compressed
text files, which may have a huge compression ratio, e.g., 25% for gzip, and 18% for
bz2. Also when scanning compressed text files, text scanners need to read entire files.
Although decreasing the rs_fetch_size_increase_factor can resolve the out of memory
issue, it will decrease the performance for other file formats as well. Thus, we'd
better have a separate factor to limit the maximum fetch size for compressed text
files only. In this patch, we introduce a new flag for this, with which no change is
required on the client side.

- Users can set the maximum fetch size for compressed text files using the flag:
rs_compressed_max_fetch_size

- If rs_compressed_max_fetch_size is a non-positve number, RecordService will not
apply it. And users can limit the maximum fetch size for each worker task on client
side when creating RecordServiceWorkerClient object.

Change-Id: Ib64de383bc4900a5293d03eecc5d732bcd998039
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/11885
Tested-by: Jenkins User
Reviewed-by: Li Li <lili@cloudera.com>
Tested-by: Li Li <lili@cloudera.com>
Reviewed-by: Chao Sun <chao.sun@cloudera.com>